Rename and refine WMF campaigns controller, add routes and tests
This is a complete first draft of the MVP API for Wikimedia Event Center (aka Campaigns Platform).

What remains is to add checks to prevent adding or removing users from an event that has syncing enabled (as the participant list will be completely derived from the Event Center participant list).